Add Visual Studio project

+++ vfsfile.cpp
@@ -5,17 +5,17 @@
 #include "vfsfile.hpp"
 #include "vfs.hpp"
 
-ssize_t VfsFile::Read(size_t bytes, void* data)
+int32 VfsFile::Read(uint32 bytes, void* data)
 {
     return 0;
 }
 
-ssize_t VfsFile::Write(size_t bytes, const void* data)
+int32 VfsFile::Write(uint32 bytes, const void* data)
 {
     return 0;
 }
 
-size_t VfsFile::Seek(ssize_t offset, VfsSeekMode mode)
+uint32 VfsFile::Seek(int32 offset, VfsSeekMode mode)
 {
     return 0;
 }
